payment-service/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
requests==2.31.0
